Container with problems
The project in a container had several issues, which were resolved with this commit. The Dockerfile was modified to copy directly to the generated JAR file. The project version was changed to a stable version. The controller and the details DTO (ProductsOrders) are now managed with a list instead of an array.

+++ billing-service/src/main/java/com/billing/service/billing_service/repositorys/ProductsOrdersRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.billing.service.billing_service.domain.ProductsOrders;
 
-public interface ProductsOrdersRepository extends JpaRepository<ProductsOrders, Integer>{
+public interface ProductsOrdersRepository extends JpaRepository<ProductsOrders, Long>{
 }
